Move cloned objects to JSON files
