+++ test/fixtures/github/no-project-columns.json
@@ -1,0 +1,16 @@
+{
+    "data": {
+      "node": {
+        "id": 2,
+        "projects": {
+          "nodes": [
+            {
+              "columns": {
+                "nodes": []
+              }
+            }
+          ]
+        }
+      }
+    }
+  }
