+++ test/project-card.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest, node */
 const projectCard = require('../data-access/project-card');
 const { mockContext } = require('./utils');
-const projectCardMoved = require('./fixtures/github/project-card-moved.json');
+const projectCardMoved = require('./fixtures/github/events/project-card-moved.json');
 const projectColumns = require('./fixtures/github/project-columns.json');
 
 describe('addProjectCard', () => {
